index.d.ts removed
