+++ packages/optimizely-sdk/lib/index.react_native.ts
@@ -30,7 +30,7 @@ import defaultEventDispatcher from './plugins/event_dispatcher/index.browser';
 import eventProcessorConfigValidator from './utils/event_processor_config_validator';
 import { createNotificationCenter } from './core/notification_center';
 import { createEventProcessor } from './plugins/event_processor';
-import { SDKOptions, OptimizelyDecideOption, Client } from './shared_types';
+import { OptimizelyDecideOption, Client, Config } from './shared_types';
 import { createHttpPollingDatafileManager } from './plugins/datafile_manager/http_polling_datafile_manager';
 
 const logger = getLogger();
@@ -47,7 +47,7 @@ const DEFAULT_EVENT_MAX_QUEUE_SIZE = 10000;
  * @return {Optimizely|null} the Optimizely object
  *                           null on error 
  */
-const createInstance = function(config: SDKOptions): Optimizely | Client | null {
+ const createInstance = function(config: Config): Client | null {
   try {
     // TODO warn about setting per instance errorHandler / logger / logLevel
     if (config.errorHandler) {
@@ -107,6 +107,7 @@ const createInstance = function(config: SDKOptions): Optimizely | Client | null 
       errorHandler,
       datafileManager:  config.sdkKey ? createHttpPollingDatafileManager(config.sdkKey, logger, config.datafile, config.datafileOptions) : undefined,
       notificationCenter,
+      isValidInstance: config.isValidInstance,
     };
 
     // If client engine is react, convert it to react native.
